requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto3==1.4.4
decorator==4.0.10
dj-database-url==0.4.0
Django==1.9.2
django-boto==0.3.9
django-filter==0.14.0
django-floppyforms==1.7.0
django-storages==1.5.1
djangorestframework==3.4.6
easy-thumbnails==2.3
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.10.1
ipython==4.2.1
ipython-genutils==0.1.0
Markdown==2.6.6
newrelic==2.78.0.57
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.3.0
psycopg2
ptyprocess==0.5.1
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wheel==0.24.0
whitenoise==2.0.6
